* release/nodefy_0.23.15: (175 commits)
  finish version++
  version++
  Stop using the room ID from the space child state.
  finish version++
  version++
  MXBeaconAggregations: Add debug log when MXBeaconInfoSummary associated to a stopped beacon info is not found.
  Implement new space selector bottom sheet (#1541)
  Rooms without an unread badge show up in "unread" tab (#1540)
  Delight edit layout experiment (#1515)
  Send an empty dictionary when calling /join to be spec compliant.
  Empty commit to wake codecov
  AllWorkingTests: Disable tests that do not pass on GH actions
  changelog
  AllWorkingTests: Disabled commented tests
  Crypto tests CI: Expose tests report on failure and share code coverage
  Tests: Run working integration tests on every PR
  AllWorkingTests: Disable all unit tests
  AllWorkingTests: Disable the 22 tests that fail
  AllWorkingTests: Disable test classes with tests that crash
  Tests: Create AllWorkingTests test plan with stable tests
  ...

# Conflicts:
#	MatrixSDK/Data/RoomSummaryStore/CoreData/Models/MXRoomSummaryMO.swift

+++ MatrixSDK/Data/Store/MXFileStore/MXFileStore.m
@@ -29,7 +29,7 @@
 #import "MatrixSDKSwiftHeader.h"
 #import "MXFileRoomSummaryStore.h"
 
-static NSUInteger const kMXFileVersion = 79;
+static NSUInteger const kMXFileVersion = 80;
 
 static NSString *const kMXFileStoreFolder = @"MXFileStore";
 static NSString *const kMXFileStoreMedaDataFile = @"MXFileStore";
@@ -902,7 +902,11 @@ static NSUInteger preloadOptions;
                 }
                 @catch (NSException *exception)
                 {
-                    MXLogError(@"[MXFileStore] Warning: MXFileRoomStore file for room %@ has been corrupted. Exception: %@", roomId, exception);
+                    NSDictionary *logDetails = @{
+                        @"roomId": roomId ?: @"",
+                        @"exception": exception
+                    };
+                    MXLogErrorWithDetails(@"[MXFileStore] Warning: MXFileRoomStore file for room has been corrupted", logDetails);
                     [self logFiles];
                     [self deleteAllData];
                 }
@@ -941,7 +945,11 @@ static NSUInteger preloadOptions;
                 }
                 @catch (NSException *exception)
                 {
-                    MXLogError(@"[MXFileStore] Warning: MXFileRoomOutgoingMessagesStore file for room %@ has been corrupted. Exception: %@", roomId, exception);
+                    NSDictionary *logDetails = @{
+                        @"roomId": roomId ?: @"",
+                        @"exception": exception
+                    };
+                    MXLogErrorWithDetails(@"[MXFileStore] Warning: MXFileRoomOutgoingMessagesStore file for room as been corrupted", logDetails);
                     [self logFiles];
                     [self deleteAllData];
                 }
@@ -981,7 +989,11 @@ static NSUInteger preloadOptions;
                 }
                 @catch (NSException *exception)
                 {
-                    MXLogError(@"[MXFileStore] Warning: loadReceipts file for room %@ has been corrupted. Exception: %@", roomId, exception);
+                    NSDictionary *logDetails = @{
+                        @"roomId": roomId ?: @"",
+                        @"exception": exception
+                    };
+                    MXLogErrorWithDetails(@"[MXFileStore] Warning: loadReceipts file for room as been corrupted", logDetails);
                     
                     // We used to reset the store and force a full initial sync but this makes the app
                     // start very slowly.
@@ -1597,11 +1609,18 @@ static NSUInteger preloadOptions;
 #pragma mark - MXFileStore metadata
 - (void)loadMetaData
 {
+    [self loadMetaData:YES];
+}
+
+- (void)loadMetaData:(BOOL)enableClearData
+{
+    MXLogDebug(@"[MXFileStore] loadMetaData: enableClearData: %@", enableClearData ? @"YES" : @"NO");
+
     NSString *metaDataFile = [storePath stringByAppendingPathComponent:kMXFileStoreMedaDataFile];
     
     @try
     {
-        metaData = [NSKeyedUnarchiver unarchiveObjectWithFile:metaDataFile];
+        metaData = [self loadRootObjectWithoutSecureCodingFromFile:metaDataFile];
     }
     @catch (NSException *exception)
     {
@@ -1623,7 +1642,10 @@ static NSUInteger preloadOptions;
     {
         MXLogDebug(@"[MXFileStore] loadMetaData: event stream token is missing");
         [self logFiles];
-        [self deleteAllData];
+        if (enableClearData)
+        {
+            [self deleteAllData];
+        }
     }
 }
 
@@ -1665,7 +1687,7 @@ static NSUInteger preloadOptions;
             self->backupEventStreamToken = self->metaData.eventStreamToken;
 
             // Store new data
-            [NSKeyedArchiver archiveRootObject:self->metaData toFile:file];
+            [self saveObject:self->metaData toFile:file];
 
 #if DEBUG
             MXLogDebug(@"[MXFileStore commit] lasted %.0fms for metadata", [[NSDate date] timeIntervalSinceDate:startDate] * 1000);
@@ -2251,9 +2273,8 @@ static NSUInteger preloadOptions;
         }
         
         id object = [NSKeyedUnarchiver unarchivedObjectOfClasses:classes fromData:data error:&error];
-        if (object)
+        if (object && !error)
         {
-            MXLogDebug(@"[MXFileStore] Loaded object from class");
             return object;
         }
         else
@@ -2297,9 +2318,8 @@ static NSUInteger preloadOptions;
         
         // Seems to be an implementation detaul
         id object = [unarchiver decodeTopLevelObjectForKey:NSKeyedArchiveRootObjectKey error:&error];
-        if (object)
+        if (object && !error)
         {
-            MXLogDebug(@"[MXFileStore] Loaded object from class");
             return object;
         }
         else
